requirements.txt
accelerate==0.18.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
blobfile==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
diffuser==0.0.1
diffusers==0.16.1
einops==0.6.1
fairscale==0.4.13
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.4.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
imageio==2.28.0
importlib-metadata==6.6.0
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
legacy==0.1.6
lxml==4.9.2
-e git+https://github.com/sail-sg/MDT.git@151e8af69432de7298002844bb289384a4440e55#egg=masked_diffusion
matplotlib==3.7.1
mmcv-full @ file:///data/home/huangshan/mmcv_full-1.7.1-cp310-cp310-manylinux1_x86_64.whl
networkx==3.1
numpy==1.24.3
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pycryptodomex==3.17
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.5.5
requests==2.29.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.13
tomli==2.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yacs==0.1.8
yapf==0.33.0
zipp==3.15.0
